+++ Hyakuninisshu/Presentations/PresentationModels/Play.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Play {
-    let no: Int
+    let no: UInt8
     let yomiFuda: YomiFuda
     let toriFudas: [ToriFuda]
     let correct: Material
